Refactor "Register catalog" code lens to be in catalog package
